pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "email-triage-env"
version = "1.0.0"
description = "OpenEnv email triage environment"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.1.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["email_triage_env*"]

[project.scripts]
server = "server.app:main"
